refactor: globa eestructure
BREAKING CHANGES:
- Removed all the route, as they are almost all useless
- Made Respositories using CrudRepositories
- Changed the package named to be plurals rather than singular
