+++ src/main/java/edu/lorsenmarek/backend/controllers/SerieController.java
@@ -1,0 +1,8 @@
+package edu.lorsenmarek.backend.controllers;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/serie")
+public class SerieController {
+
+}
