+++ src/main/java/edu/lorsenmarek/backend/models/User.java
@@ -1,13 +1,13 @@
-package edu.lorsenmarek.backend.model;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+package edu.lorsenmarek.backend.models;
+import lombok.*;
+import org.springframework.data.annotation.Id;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class User {
+    @Id
     private Long id;
     private String firstName;
     private String lastName;
