+++ src/main/java/edu/lorsenmarek/backend/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package edu.lorsenmarek.backend.repositories;
+
+import edu.lorsenmarek.backend.models.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+}
